Allow user to set simulation outputs
Modify the script to use the method developed in the ResStock to EnergyPlus pipeline for changing the flags in the IDF files that tell EnergyPlus what simulations outputs to save.

Change eprun_s.py to check if user supplied output requests. Then load the scripts generated for the upstream workflow and set the energyplus simulation output.

Add some optional arguments to the script_epwork.py to accomodate changes.